requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosmtplib==3.0.2
altgraph==0.17.4
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.7.4
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
packaging==24.1
pefile==2023.2.7
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
typing_extensions==4.12.2
yarl==1.9.4
requests
pandas